# openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "API de Líneas de Microbuses",
        "version": "1.0.0",
        "description": "Documentación de líneas, rutas y puntos (ida/vuelta)."
    },
    "paths": {
        "/api/lineas/": {
            "get": {
                "operationId": "listar_lineas",
                "description": "Devuelve la lista completa de líneas de microbús.",
                "summary": "Listar todas las líneas",
                "tags": [
                    "lineas"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {
                        "basicAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Linea"
                                    }
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/lineas/{codigo}/rutas/{sentido}/puntos/": {
            "get": {
                "operationId": "obtener_puntos_ruta_linea",
                "description": "Dado el código de una línea (ej. L001) y el sentido de la ruta (1 = ida, 2 = vuelta), devuelve todos los puntos ordenados con:\n- tiempo del tramo\n- tiempo acumulado hasta cada punto (tiempo de llegada)\n- distancia del tramo\n- distancia acumulada",
                "summary": "Puntos de una línea por sentido (ida/vuelta)",
                "parameters": [
                    {
                        "in": "path",
                        "name": "codigo",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Código de la línea (por ejemplo L001)",
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "sentido",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Sentido de la ruta: 1 = ida, 2 = vuelta",
                        "required": true
                    }
                ],
                "tags": [
                    "lineas"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {
                        "basicAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RutaConPuntos"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Linea": {
                "type": "object",
                "description": "Serializer simple para listar líneas.\nDevuelve: id, codigo, nombre y color (hex).",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "codigo": {
                        "type": "string",
                        "maxLength": 10
                    },
                    "nombre": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "color": {
                        "type": "string",
                        "maxLength": 7
                    }
                },
                "required": [
                    "codigo",
                    "color",
                    "id"
                ]
            },
...
            "PuntoRuta": {
                "type": "object",
                "properties": {
                    "orden": {
                        "type": "integer"
                    },
                    "latitud": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,3}(?:\\.\\d{0,6})?$"
                    },
                    "longitud": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,3}(?:\\.\\d{0,6})?$"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "distancia_tramo": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,8}(?:\\.\\d{0,2})?$"
                    },
                    "distancia_acumulada": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,8}(?:\\.\\d{0,2})?$"
                    },
                    "tiempo_tramo": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,4}(?:\\.\\d{0,2})?$"
                    },
                    "tiempo_acumulado": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,4}(?:\\.\\d{0,2})?$"
                    }
                },
                "required": [
                    "descripcion",
                    "distancia_acumulada",
                    "distancia_tramo",
                    "latitud",
                    "longitud",
                    "orden",
                    "tiempo_acumulado",
                    "tiempo_tramo"
                ]
            },
...
            "RutaConPuntos": {
                "type": "object",
                "properties": {
                    "linea_codigo": {
                        "type": "string"
                    },
                    "linea_nombre": {
                        "type": "string"
                    },
                    "sentido": {
                        "type": "integer"
                    },
                    "descripcion_ruta": {
                        "type": "string"
                    },
                    "distancia_total": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,8}(?:\\.\\d{0,2})?$"
                    },
                    "tiempo_total": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,4}(?:\\.\\d{0,2})?$"
                    },
                    "puntos": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PuntoRuta"
                        }
                    }
                },
                "required": [
                    "descripcion_ruta",
                    "distancia_total",
                    "linea_codigo",
                    "linea_nombre",
                    "puntos",
                    "sentido",
                    "tiempo_total"
                ]
            },
...
        },
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            },
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            }
        }
    }
}